requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.3
anyio==3.6.2
apache-airflow==2.6.1
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-google==10.0.0
apache-airflow-providers-http==4.3.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-snowflake==4.0.5
apache-airflow-providers-sqlite==3.3.1
apispec==3.3.2
argcomplete==3.0.5
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
backoff==2.2.1
beautifulsoup4==4.12.0
blinker==1.6
cachelib==0.9.0
cachetools==5.3.0
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.2.35
croniter==1.3.10
cryptography==40.0.1
db-dtypes==1.1.1
Deprecated==1.2.13
dill==0.3.6
dnspython==2.3.0
docutils==0.19
email-validator==1.3.1
exceptiongroup==1.1.1
filelock==3.10.7
Flask==2.2.3
Flask-AppBuilder==4.3.0
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.3.3
future==0.18.3
gcloud-aio-auth==4.2.1
gcloud-aio-bigquery==6.3.0
gcloud-aio-storage==8.2.0
google-api-core==2.8.2
google-api-python-client==1.12.11
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-cloud-aiplatform==1.16.1
google-cloud-appengine-logging==1.1.3
google-cloud-audit-log==0.2.4
google-cloud-automl==2.8.0
google-cloud-bigquery==2.34.4
google-cloud-bigquery-datatransfer==3.7.0
google-cloud-bigquery-storage==2.14.1
google-cloud-bigtable==2.11.1
google-cloud-build==3.9.0
google-cloud-compute==0.7.0
google-cloud-container==2.11.1
google-cloud-core==2.3.2
google-cloud-datacatalog==3.9.0
google-cloud-dataflow-client==0.5.4
google-cloud-dataform==0.2.0
google-cloud-dataplex==1.1.0
google-cloud-dataproc==5.0.0
google-cloud-dataproc-metastore==1.6.0
google-cloud-dlp==3.8.0
google-cloud-kms==2.12.0
google-cloud-language==1.3.2
google-cloud-logging==3.2.1
google-cloud-memcache==1.4.1
google-cloud-monitoring==2.11.0
google-cloud-orchestration-airflow==1.4.1
google-cloud-os-login==2.7.1
google-cloud-pubsub==2.13.5
google-cloud-redis==2.9.0
google-cloud-resource-manager==1.6.0
google-cloud-secret-manager==1.0.2
google-cloud-spanner==1.19.3
google-cloud-speech==1.3.4
google-cloud-storage==2.8.0
google-cloud-tasks==2.10.1
google-cloud-texttospeech==1.0.3
google-cloud-translate==1.7.2
google-cloud-videointelligence==1.16.3
google-cloud-vision==1.0.2
google-cloud-workflows==1.7.1
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.56.4
graphviz==0.20.1
greenlet==2.0.2
grpc-google-iam-v1==0.12.4
grpcio==1.54.0
grpcio-gcp==0.2.2
grpcio-status==1.48.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
json-merge-patch==0.2
jsonschema==4.17.3
lazy-object-proxy==1.9.0
limits==3.5.0
linkify-it-py==2.0.0
lockfile==0.12.2
looker-sdk==23.8.1
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.5
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.3
oauthlib==3.2.2
ordered-set==4.1.0
oscrypto==1.3.0
packaging==21.3
pandas==2.0.1
pandas-gbq==0.17.9
pathspec==0.9.0
pendulum==2.1.2
pluggy==1.0.0
prison==0.2.1
proto-plus==1.19.6
protobuf==3.20.0
psutil==5.9.4
py4j==0.10.9.5
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.7
pydata-google-auth==1.8.0
Pygments==2.14.0
PyJWT==2.6.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.19.3
pyspark==3.3.2
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.3.3
rich_argparse==1.1.0
rsa==4.9
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
snowflake-connector-python==3.0.2
snowflake-sqlalchemy==1.4.7
soupsieve==2.4
SQLAlchemy==1.4.47
sqlalchemy-bigquery==1.6.1
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.40.0
sqlparse==0.4.3
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.2.0
text-unidecode==1.3
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.8.2
